# openapi.yaml
openapi: 3.0.3
info:
  title: Sports Companion API
  version: 1.0.0
servers:
  - url: https://api.sportification.app/v1
paths:
...
  /chats:
    get:
      summary: List chats
      responses:
        '200':
          description: Successful response
    post:
      summary: Create chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Chat'
      responses:
        '201':
          description: Chat created
  /chats/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      summary: Get chat
      responses:
        '200':
          description: Chat data
    patch:
      summary: Update chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Chat'
      responses:
        '200':
          description: Chat updated
...
  /messages/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      summary: Get message
      responses:
        '200':
          description: Message data
    patch:
      summary: Update message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Message'
      responses:
        '200':
          description: Message updated
...
components:
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema:
        type: string
  schemas:
...
    Chat:
      type: object
      properties:
        id: { type: string }
        participants:
          type: array
          items: { type: string }
        messages:
          type: array
          items: { type: string }
        media:
          type: array
          items: { type: string }
        reactions: { type: object }
        threads:
          type: array
          items: { type: string }
...
    Message:
      type: object
      properties:
        id: { type: string }
        sender: { type: string }
        content: { type: string }
        timestamp:
          type: string
          format: date-time
        media:
          type: array
          items: { type: string }
        reactions: { type: object }
        thread: { type: string }
...